Dashboard refactor
Changed base to run via streamlit rather than on Dash under a new version. Removed prior files to reflect changes to new source. Added new requirements.txt and moved previous version to a new, deprecated folder
