+++ .gitignore
@@ -1,0 +1,9 @@
+test/test.ipynb
+data/archived/clean_fall.csv
+data/archived/clean_sping.csv
+data/archived/FA23-standardized.xlsx
+data/archived/FA23.csv
+data/archived/Fall 2023.csv
+data/archived/SP23 Standardized.xlsx
+data/archived/SP23.csv
+data/archived/Spring 2023.csv
