+++ deprecatedVers/v1.0/requirements.txt
@@ -1,0 +1,6 @@
+dash==2.14.2
+pandas==1.5.3
+numpy
+gunicorn
+openpyxl
+dash-bootstrap-components
