+++ requirements.txt
@@ -1,6 +1,5 @@
-dash==2.14.2
-pandas==1.5.3
 numpy
+pandas
+streamlit
 gunicorn
 openpyxl
-dash-bootstrap-components
